Refactor PersonSearchFields into smaller components;
